feat(onboarding): enhance application form page 5 with competency questions and submission flow, update modals and localization

+++ main/src/mongoose/schemas/onboardingTrackerSchema.ts
@@ -1,8 +1,5 @@
 import mongoose, { Schema } from "mongoose";
-import {
-  EStepPath,
-  IOnboardingTrackerDoc,
-} from "@/types/onboardingTracker.type";
+import { EStepPath, IOnboardingTrackerDoc } from "@/types/onboardingTracker.type";
 import { ECompanyApplicationType } from "@/hooks/frontendHooks/useCompanySelection";
 import PreQualifications from "../models/Prequalifications";
 import ApplicationForm from "../models/applicationForm";
@@ -30,9 +27,7 @@ const onboardingTrackerSchema = new Schema<IOnboardingTrackerDoc>(
       type: String,
       enum: {
         values: Object.values(ECompanyApplicationType),
-        message: `application type can only be one of ${Object.values(
-          ECompanyApplicationType
-        )}`,
+        message: `application type can only be one of ${Object.values(ECompanyApplicationType)}`,
       },
     },
     status: {
@@ -40,9 +35,7 @@ const onboardingTrackerSchema = new Schema<IOnboardingTrackerDoc>(
         type: String,
         enum: {
           values: Object.values(EStepPath),
-          message: `Current step must be one of: ${Object.values(
-            EStepPath
-          ).join(", ")}`,
+          message: `Current step must be one of: ${Object.values(EStepPath).join(", ")}`,
         },
         required: [true, "Current onboarding step is required."],
       },
@@ -50,9 +43,7 @@ const onboardingTrackerSchema = new Schema<IOnboardingTrackerDoc>(
         type: String,
         enum: {
           values: Object.values(EStepPath),
-          message: `Completed step must be one of: ${Object.values(
-            EStepPath
-          ).join(", ")}`,
+          message: `Completed step must be one of: ${Object.values(EStepPath).join(", ")}`,
         },
         required: [true, "Completed onboarding step is required."],
       },
@@ -66,9 +57,7 @@ const onboardingTrackerSchema = new Schema<IOnboardingTrackerDoc>(
       required: [true, "Company id is requried"],
       enum: {
         values: Object.values(ECompanyId),
-        message: `application type can only be one of ${Object.values(
-          ECompanyId
-        )}`,
+        message: `company id can only be one of ${Object.values(ECompanyId)}`,
       },
     },
     forms: {
@@ -129,32 +118,19 @@ onboardingTrackerSchema.post("findOneAndDelete", async function (doc) {
   const { forms } = doc;
   const tasks = [];
 
-  if (forms.preQualification)
-    tasks.push(PreQualifications.deleteOne({ _id: forms.preQualification }));
+  if (forms.preQualification) tasks.push(PreQualifications.deleteOne({ _id: forms.preQualification }));
 
-  if (forms.driverApplication)
-    tasks.push(ApplicationForm.deleteOne({ _id: forms.driverApplication }));
+  if (forms.driverApplication) tasks.push(ApplicationForm.deleteOne({ _id: forms.driverApplication }));
 
-  if (forms.consents)
-    tasks.push(PoliciesConsents.deleteOne({ _id: forms.consents }));
+  if (forms.consents) tasks.push(PoliciesConsents.deleteOne({ _id: forms.consents }));
 
-  if (forms.carrierEdge)
-    tasks.push(
-      mongoose.model("CarrierEdgeStatus").deleteOne({ _id: forms.carrierEdge })
-    );
+  if (forms.carrierEdge) tasks.push(mongoose.model("CarrierEdgeStatus").deleteOne({ _id: forms.carrierEdge }));
 
-  if (forms.driveTest)
-    tasks.push(mongoose.model("DriveTest").deleteOne({ _id: forms.driveTest }));
+  if (forms.driveTest) tasks.push(mongoose.model("DriveTest").deleteOne({ _id: forms.driveTest }));
 
-  if (forms.drugTest)
-    tasks.push(mongoose.model("DrugTest").deleteOne({ _id: forms.drugTest }));
+  if (forms.drugTest) tasks.push(mongoose.model("DrugTest").deleteOne({ _id: forms.drugTest }));
 
-  if (forms.flatbedTraining)
-    tasks.push(
-      mongoose
-        .model("FlatbedTraining")
-        .deleteOne({ _id: forms.flatbedTraining })
-    );
+  if (forms.flatbedTraining) tasks.push(mongoose.model("FlatbedTraining").deleteOne({ _id: forms.flatbedTraining }));
 
   await Promise.all(tasks);
 });
